refactor: move mergeThemeWithBrandConfig utility to utils directory and remove unused stories
- Moved the `mergeThemeWithBrandConfig` function to `utils/merge_theme_with_brand_config.js`.
- Deleted the associated story file `merge_theme_with_brand_config.stories.tsx` as it was no longer needed.
