+++ src/utils/merge_theme_with_brand_config.stories.tsx
@@ -8,7 +8,7 @@ import { mergeThemeWithBrandConfig } from './merge_theme_with_brand_config.js'
 const TestComponent = () => <div>Testing mergeThemeWithBrandConfig function</div>
 
 const meta: Meta<typeof TestComponent> = {
-	title: 'Utils/mergeThemeWithBrandConfig',
+	title: 'utils/mergeThemeWithBrandConfig',
 	tags: ['unit'],
 	component: TestComponent,
 	parameters: {
